Google translate refactoring
