+++ backend/core/src/listings/listings.service.ts
@@ -16,36 +16,15 @@ import { plainToClass } from "class-transformer"
 import { PropertyCreateDto, PropertyUpdateDto } from "../property/dto/property.dto"
 import { arrayIndex } from "../libs/arrayLib"
 import { addFilters } from "../shared/filter"
-import { Translate } from "@google-cloud/translate/build/src/v2"
 import { Language } from "../../types"
-
-const TRANSLATION_KEYS = [
-  "applicationPickUpAddressOfficeHours",
-  "costsNotIncluded",
-  "creditHistory",
-  "criminalBackground",
-  "programRules",
-  "rentalAssistance",
-  "rentalHistory",
-  "requiredDocuments",
-  "specialNotes",
-  {
-    property: [
-      "accessibility",
-      "amenities",
-      "petPolicy",
-      "servicesOffered",
-      "smokingPolicy",
-      "unitAmenities",
-    ],
-    whatToExpect: ["applicantsWillBeContacted", "allInfoWillBeVerified", "bePreparedIfChosen"],
-    preferences: ["title", "description"],
-  },
-]
+import { TranslationsService } from "../translations/translations.service"
 
 @Injectable()
 export class ListingsService {
-  constructor(@InjectRepository(Listing) private readonly listingRepository: Repository<Listing>) {}
+  constructor(
+    @InjectRepository(Listing) private readonly listingRepository: Repository<Listing>,
+    private readonly translationService: TranslationsService
+  ) {}
 
   private getFullyJoinedQueryBuilder() {
     return this.listingRepository
@@ -226,70 +205,9 @@ export class ListingsService {
       throw new NotFoundException()
     }
     if (lang !== Language.en) {
-      // Get key-value pairs from listing to be translated
-      const translations = this.getTranslations(result)
-
-      const translatedValues = await this.translateService().translate(translations.values, {
-        from: Language.en,
-        to: lang,
-      })
-      // Attach translated values to the listing
-      translations.keys.forEach((key, index) => {
-        this.setValue(result, key, translatedValues[0][index])
-      })
+      await this.translationService.translateListing(result, lang)
     }
 
-    return result
-  }
-
-  private translateService = () => {
-    return new Translate({
-      credentials: {
-        private_key: process.env.GOOGLE_API_KEY.replace(/\\n/gm, "\n"),
-        client_email: process.env.GOOGLE_API_EMAIL,
-      },
-      projectId: process.env.GOOGLE_API_ID,
-    })
-  }
-  // Sets value to the object by string path. eg. "property.accessibility" or "preferences.0.title"
-  private setValue = (object, path, value) =>
-    path
-      .split(".")
-      .reduce(
-        (currentObject, currentPath, index) =>
-          (currentObject[currentPath] =
-            path.split(".").length === ++index ? value : currentObject[currentPath] || {}),
-        object
-      )
-
-  // Returns not null key-values pairs also from nested properties
-  private findData = (keys, object, results, parent = null) => {
-    keys.forEach((key) => {
-      if (typeof key === "string") {
-        if (Array.isArray(object)) {
-          object.forEach((value, i) => {
-            results.keys.push(parent ? [parent, i, key].join(".") : key)
-            results.values.push(value[key])
-          })
-        } else {
-          if (object[key]) {
-            results.keys.push(parent ? [parent, key].join(".") : key)
-            results.values.push(object[key])
-          }
-        }
-        return
-      } else {
-        for (const k in key) {
-          if (object[k]) {
-            this.findData(key[k], object[k], results, k)
-          }
-        }
-      }
-    })
-  }
-  private getTranslations = (listing: Listing) => {
-    const result = { keys: [], values: [] }
-    this.findData(TRANSLATION_KEYS, listing, result)
     return result
   }
 }
